+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+### Changed
+
+ - Permissions fixer now works with Magento 2.2+
+
 ### Added
 
 - Build script added for images